backend/requirements.txt
fastapi
uvicorn
gunicorn
pymongo
python-dotenv
passlib[bcrypt]
pydantic[email]
google-cloud-logging
websockets
